fix ignition compatibility - #711 #2025
